refactor: circleとtriangleを一つのファイルに
