+++ Makefile
@@ -11,8 +11,7 @@ src =\
 	./data.c \
 	./term3d.c \
 	./draw.c \
-	./circle.c \
-	./triangle.c \
+	./calc.c \
 	./camera.c \
 	./move.c \
 	./reader.c \
@@ -50,8 +49,7 @@ testdir = ./gtest
 srcs_test = \
 	./$(src_dir)/draw.c \
 	./$(src_dir)/data.c \
-	./$(src_dir)/circle.c \
-	./$(src_dir)/triangle.c \
+	./$(src_dir)/calc.c \
 	./$(src_dir)/move.c \
 	./$(src_dir)/reader.c \
 	./$(src_dir)/camera.c \
