+++ gtest/gtest_calc.cpp
@@ -1,7 +1,7 @@
 #include <gtest/gtest.h>
-#include "circle.h"
+#include "calc.h"
 
-TEST(Circle, intersect_with_plane)
+TEST(Calc, intersect_with_plane)
 {
 	// カメラが平面を向いている
   t_camera  *camera;
@@ -70,7 +70,7 @@ TEST(Circle, intersect_with_plane)
 	free(camera);
 }
 
-TEST(Circle, is_intersect_with_circle)
+TEST(Calc, is_intersect_with_circle)
 {
 	t_circle *circle = (t_circle *)malloc(sizeof(t_circle));
 	circle->center = vect_new(0, 0, 0);
@@ -100,4 +100,21 @@ TEST(Circle, is_intersect_with_circle)
 
 	free(camera);
 	free(circle);
+}
+
+TEST(Calc, is_intersect_with_triangle)
+{
+	t_triangle triangle;
+
+	triangle.vert1 = vect_new(0, 2, 0);
+	triangle.vert2 = vect_new(-2, -2, 0);
+	triangle.vert3 = vect_new(2, -2, 0);
+
+	t_camera *camera = (t_camera *)malloc(sizeof(t_camera));
+	camera->pos = vect_new(0, 0, 15);
+	camera->ray = vect_new(0, 0, -1);
+	EXPECT_EQ(is_intersect_with_triangle(camera, &triangle), true);
+	// EXPECT_EQ(is_equal(vect_distance(camera->lookat, camera->pos), 15), true);
+  // 中心点がないのでいい感じの距離がつかめないためコメントアウト。概数を目視チェックしていた。
+	// EXPECT_DOUBLE_EQ(vect_distance(camera->lookat, triangle.vect1), 15);
 }
