+++ includes/data.h
@@ -57,6 +57,7 @@ typedef struct s_data
 {
 	t_camera	*camera;
 	t_circle	circle[1000];
+	t_triangle	triangle[1000];
 	int			count;
 	t_type		type;
 }				t_data;
